pyproject.toml
[tool.poetry]
name = "fastapi-fargate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
